tech(REC-2) : move routes and app in src
