+++ server/app.js
@@ -5,9 +5,9 @@ const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 const cors = require('cors');
 
-const articles = require('./routes/articles');
-const index = require('./routes/index');
-const users = require('./routes/users');
+const articles = require('./src/routes/articles');
+const index = require('./src/routes/index');
+const users = require('./src/routes/users');
 
 const app = express();
